refactor : 카테고리 및 예외 구문 수정(Test)
