+++ .env.example
@@ -1,6 +1,11 @@
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=jinjuwiki
+DB_USERNAME=your-db-username
+DB_PASSWORD=your-db-password
 MAIL_HOST=smtp.gmail.com
 MAIL_PORT=587
 MAIL_USERNAME=your-email@gmail.com
 MAIL_PASSWORD=your-16-digit-app-password
 MAIL_FROM=your-email@gmail.com
-JWT_SECRET=jinjuwiki-jwt-secret-key-for-local-development-2026
+JWT_SECRET=your-jwt-secret
