+++ src/test/java/com/jinju/jinjuwiki/domain/auth/service/AuthServiceTest.java
@@ -3,12 +3,12 @@ package com.jinju.jinjuwiki.domain.auth.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationSendRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationVerifyRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.LoginRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.LoginResponse;
-import com.jinju.jinjuwiki.domain.auth.dto.SignupRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.SignupResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationSendRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationVerifyRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.request.LoginRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.LoginResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.SignupRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.SignupResponse;
 import com.jinju.jinjuwiki.domain.auth.repository.EmailVerificationRepository;
 import com.jinju.jinjuwiki.global.error.BusinessException;
 import com.jinju.jinjuwiki.global.error.ErrorCode;
