+++ src/test/java/com/jinju/jinjuwiki/domain/document/service/DocumentServiceTest.java
@@ -1,12 +1,12 @@
 package com.jinju.jinjuwiki.domain.document.service;
 
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationSendRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.EmailVerificationVerifyRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationSendRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.request.EmailVerificationVerifyRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.jinju.jinjuwiki.domain.auth.dto.SignupRequest;
-import com.jinju.jinjuwiki.domain.auth.dto.SignupResponse;
+import com.jinju.jinjuwiki.domain.auth.dto.request.SignupRequest;
+import com.jinju.jinjuwiki.domain.auth.dto.response.SignupResponse;
 import com.jinju.jinjuwiki.domain.auth.repository.EmailVerificationRepository;
 import com.jinju.jinjuwiki.domain.auth.service.AuthService;
 import com.jinju.jinjuwiki.domain.auth.service.EmailSender;
@@ -33,6 +33,12 @@ import com.jinju.jinjuwiki.domain.document.repository.DocumentRepository;
 @Transactional
 @SpringBootTest
 class DocumentServiceTest {
+
+    private static final String SCHOOL = "학교";
+    private static final String STUDENT = "학생";
+    private static final String INCIDENT = "사건사고";
+    private static final String TEACHER = "선생님";
+    private static final String ETC = "기타";
 
     @Autowired
     private DocumentService documentService;
@@ -65,7 +71,7 @@ class DocumentServiceTest {
     void createDocumentSuccess() {
         verifyEmail("doc1@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc1@test.com", "password123", "docUser"));
-        Category category = categoryRepository.findByName("학교").orElseThrow();
+        Category category = getCategory(SCHOOL);
 
         DocumentCreateResponse response = documentService.createDocument(
                 new DocumentCreateRequest("문서 제목", "문서 본문", category.getId()),
@@ -80,7 +86,7 @@ class DocumentServiceTest {
     @Test
     @DisplayName("존재하지 않는 작성자 ID로 문서를 생성하면 예외가 발생한다.")
     void createDocumentFailWhenUserNotFound() {
-        Category category = categoryRepository.findByName("학교").orElseThrow();
+        Category category = getCategory(SCHOOL);
 
         assertThatThrownBy(() -> documentService.createDocument(
                 new DocumentCreateRequest("문서 제목", "문서 본문", category.getId()),
@@ -96,7 +102,7 @@ class DocumentServiceTest {
     void getDocumentSuccess() {
         verifyEmail("doc2@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc2@test.com", "password123", "docUser2"));
-        Category category = categoryRepository.findByName("학생").orElseThrow();
+        Category category = getCategory(STUDENT);
         DocumentCreateResponse created = documentService.createDocument(
                 new DocumentCreateRequest("수학 공부법", "개념부터 반복", category.getId()),
                 user.userId()
@@ -113,7 +119,7 @@ class DocumentServiceTest {
     void getDocumentsSuccess() {
         verifyEmail("doc3@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc3@test.com", "password123", "docUser3"));
-        Category category = categoryRepository.findByName("사건사고").orElseThrow();
+        Category category = getCategory(INCIDENT);
 
         documentService.createDocument(new DocumentCreateRequest("첫 번째 글", "내용 1", category.getId()), user.userId());
         documentService.createDocument(new DocumentCreateRequest("두 번째 글", "내용 2", category.getId()), user.userId());
@@ -129,8 +135,8 @@ class DocumentServiceTest {
     void updateDocumentSuccess() {
         verifyEmail("doc4@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc4@test.com", "password123", "docUser4"));
-        Category category = categoryRepository.findByName("학교").orElseThrow();
-        Category updatedCategory = categoryRepository.findByName("선생님").orElseThrow();
+        Category category = getCategory(SCHOOL);
+        Category updatedCategory = getCategory(TEACHER);
         DocumentCreateResponse created = documentService.createDocument(
                 new DocumentCreateRequest("원래 제목", "원래 본문", category.getId()),
                 user.userId()
@@ -144,7 +150,7 @@ class DocumentServiceTest {
 
         assertThat(response.title()).isEqualTo("수정 제목");
         assertThat(response.content()).isEqualTo("수정 본문");
-        assertThat(response.categoryName()).isEqualTo("선생님");
+        assertThat(response.categoryName()).isEqualTo(TEACHER);
     }
 
     @Test
@@ -154,7 +160,7 @@ class DocumentServiceTest {
         SignupResponse author = authService.signup(new SignupRequest("doc5@test.com", "password123", "author5"));
         verifyEmail("doc6@test.com");
         SignupResponse otherUser = authService.signup(new SignupRequest("doc6@test.com", "password123", "other6"));
-        Category category = categoryRepository.findByName("학교").orElseThrow();
+        Category category = getCategory(SCHOOL);
         DocumentCreateResponse created = documentService.createDocument(
                 new DocumentCreateRequest("제목", "본문", category.getId()),
                 author.userId()
@@ -175,7 +181,7 @@ class DocumentServiceTest {
     void deleteDocumentSuccess() {
         verifyEmail("doc7@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc7@test.com", "password123", "docUser7"));
-        Category category = categoryRepository.findByName("기타").orElseThrow();
+        Category category = getCategory(ETC);
         DocumentCreateResponse created = documentService.createDocument(
                 new DocumentCreateRequest("삭제 제목", "삭제 본문", category.getId()),
                 user.userId()
@@ -193,7 +199,7 @@ class DocumentServiceTest {
         SignupResponse author = authService.signup(new SignupRequest("doc8@test.com", "password123", "author8"));
         verifyEmail("doc9@test.com");
         SignupResponse otherUser = authService.signup(new SignupRequest("doc9@test.com", "password123", "other9"));
-        Category category = categoryRepository.findByName("학교").orElseThrow();
+        Category category = getCategory(SCHOOL);
         DocumentCreateResponse created = documentService.createDocument(
                 new DocumentCreateRequest("삭제 전 제목", "삭제 전 본문", category.getId()),
                 author.userId()
@@ -210,8 +216,8 @@ class DocumentServiceTest {
     void searchDocumentsSuccess() {
         verifyEmail("doc10@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc10@test.com", "password123", "docUser10"));
-        Category schoolCategory = categoryRepository.findByName("학교").orElseThrow();
-        Category studyCategory = categoryRepository.findByName("학생").orElseThrow();
+        Category schoolCategory = getCategory(SCHOOL);
+        Category studyCategory = getCategory(STUDENT);
 
         documentService.createDocument(
                 new DocumentCreateRequest("진주 학교 생활", "학교 행사 정보", schoolCategory.getId()),
@@ -233,8 +239,8 @@ class DocumentServiceTest {
     void searchDocumentsWithCategory() {
         verifyEmail("doc11@test.com");
         SignupResponse user = authService.signup(new SignupRequest("doc11@test.com", "password123", "docUser11"));
-        Category schoolCategory = categoryRepository.findByName("학교").orElseThrow();
-        Category studyCategory = categoryRepository.findByName("학생").orElseThrow();
+        Category schoolCategory = getCategory(SCHOOL);
+        Category studyCategory = getCategory(STUDENT);
 
         documentService.createDocument(
                 new DocumentCreateRequest("학교 시험 정보", "학교 공지", schoolCategory.getId()),
@@ -249,7 +255,7 @@ class DocumentServiceTest {
                 documentService.searchDocuments("시험", studyCategory.getId(), 0, 10);
 
         assertThat(response.content()).hasSize(1);
-        assertThat(response.content().get(0).categoryName()).isEqualTo("학생");
+        assertThat(response.content().get(0).categoryName()).isEqualTo(STUDENT);
     }
 
     @Test
@@ -265,5 +271,10 @@ class DocumentServiceTest {
         authService.sendVerificationCode(new EmailVerificationSendRequest(email));
         String code = emailVerificationRepository.findByEmail(email).orElseThrow().getCode();
         authService.verifyCode(new EmailVerificationVerifyRequest(email, code));
+    }
+
+    private Category getCategory(String categoryName) {
+        return categoryRepository.findByName(categoryName)
+                .orElseThrow(() -> new AssertionError("테스트용 카테고리를 찾을 수 없습니다: " + categoryName));
     }
 }
